# Conflicts:
#	presentation/src/main/java/com/woowahan/banchan/ui/viewmodel/OrderListViewModel.kt
#	presentation/src/main/java/com/woowahan/banchan/ui/viewmodel/RecentViewedViewModel.kt

+++ data/src/main/java/com/woowahan/data/entity/dto/RecentViewedDto.kt
@@ -15,7 +15,7 @@ data class RecentViewedDto(
     var recentViewedItemInfo: BanchanItemTableEntity
 ) {
     fun toEntity(): RecentViewedEntity = RecentViewedEntity(
-        recentViewed.id,
+        recentViewed.id.toLong(),
         recentViewedItemInfo.hash,
         recentViewed.time,
         recentViewedItemInfo.title
